Handle errors while fetching pool data more gracefully.
If we can't determine the number of shares a given pool has in its current
round, we'll simply assume twice the difficulty. This is an arbitrary number
that makes it so the pool will probably not be selected, but minimizes the
chance of incorrectly selecting it.

In any case, if we can't fetch this information, the pool may very well be
completely down anyway.

Signed-off-by: Jason Lynch <jason@calindora.com>