Assume hoppers provide a 30GH/s boost to a pool's hash rate.
The problem here is that by using the average hash rate to estimate shares, we
will always hop later than we should, because a pool's hash rate is higher prior
to the 43.5% mark, due to other hoppers. This causes us to lowball the shares
estimate. This adjustment will help to mitigate that problem, but it may be too
low.

Signed-off-by: Jason Lynch <jason@calindora.com>