Add support for donation percentages.
This allows users to add another field in pools.conf, specifying their donation
percentage to that pool. This allows the miner to calculate a more accurate
utility of the pool. In addition, by fudging the numbers, the user can promote a
pool upward even if its utility is slightly lower.

Signed-off-by: Jason Lynch <jason@calindora.com>